refactor: vat.d.ts added
