+++ types/windev/society_2.d.ts
@@ -1,9 +1,28 @@
 // Import Internal Dependencies
-import { Account } from "./account";
-import { PaymentDeadline } from "./account";
+import {
+  Account,
+  PaymentDeadline,
+  PaymentType,
+  RIB
+} from "./account";
 
-// Import Third-party Dependencies
-import { Except } from "type-fest";
+import {
+  VatParam,
+  VatType,
+  VatExigility,
+  VatRegime
+} from "./vat";
+
+export {
+  Account,
+  PaymentDeadline,
+  PaymentType,
+  VatParam,
+  VatExigility,
+  VatType,
+  VatRegime,
+  RIB
+}
 
 interface CommonField {
   id: number;
@@ -586,51 +605,6 @@ export interface FiscalFile {
   };
 
   society_status: FiscalFileGenericField;
-}
-
-export interface RIB {  //  Déplacer dans ./account.d.ts ???
-  rib_id: number;
-  diary_id: number;
-  diary_label: string;
-  society_id: number;
-  start_date: string;
-  iban: string;
-  bic: string;
-  is_default: boolean;
-}
-
-// Utilisé dans ./account.d.ts uniquement.
-// Pas trouvé de modèle sur Postman.
-// Pas trouvé de référence sur Confluence ni sur la doc partenaires.
-// vérifier que vat_type === VatType (GET/vat_param/vat_type)
-export interface VatParam { 
-  vat_param_id: number;
-  code: string;
-  account_ded: Except<Account, "counterpart_account">;
-  account_coll: Except<Account, "counterpart_account">;
-  vat: {
-    id: number;
-    rate: number;
-  };
-  vat_type: VatType;
-  vat_exigility: VatExigility;
-  blocked: boolean;
-}
-
-export interface VatType {
-  id: number;
-  label: string;
-}
-
-export interface VatExigility {
-  id: number;
-  label: string;
-}
-
-export interface VatRegime {
-  id: number;
-  name: string;
-  code: string;
 }
 
 export interface RegimeImpot {
